+++ examples/02_cook_book/save_and_load/load_03.py
@@ -1,0 +1,10 @@
+
+
+from flock.core import FlockAgent
+from flock.core.logging.formatters.pprint_formatter import PrettyPrintFormatter
+
+# chatty is a rather compley agent. no problem for Flock to load it.
+loaded_bloggy = FlockAgent.load_from_file("examples/data/chatty.json")
+
+result = loaded_bloggy.run(inputs={})
+PrettyPrintFormatter().display_data(result)
