+++ examples/data/custom_bloggy.json
@@ -1,0 +1,1 @@
+{"name": "custom_bloggy", "model": "openai/gpt-4o", "description": "", "input": "blog_idea", "output": "result", "tools": null, "use_cache": true, "hand_off": null, "termination": null, "config": {"agent_type_override": null, "disable_output": false, "temperature": 0.0, "max_tokens": 2000}, "initialize_callback": null, "evaluate_callback": "80059593020000000000008c17636c6f75647069636b6c652e636c6f75647069636b6c65948c0e5f6d616b655f66756e6374696f6e9493942868008c0d5f6275696c74696e5f747970659493948c08436f6465547970659485945294284b024b004b004b024b024b8343124b00010097006401640269015300ad037701944e8c06726573756c74948c0d48656c6c6f2c20776f726c6421948794298c0473656c66948c06696e707574739486948c412f55736572732f617261747a2f666c6f636b2f6578616d706c65732f30325f636f6f6b5f626f6f6b2f736176655f616e645f6c6f61642f736176655f30322e7079948c086576616c75617465948c086576616c75617465944b0a4311e800f88000d80c14906fd00b26d00426f994430482050701942929749452947d94288c0b5f5f7061636b6167655f5f948c00948c085f5f6e616d655f5f948c085f5f6d61696e5f5f948c085f5f66696c655f5f94680f754e4e4e7494529468008c125f66756e6374696f6e5f7365747374617465949394681d7d947d942868198c086576616c75617465948c0c5f5f7175616c6e616d655f5f948c086576616c75617465948c0f5f5f616e6e6f746174696f6e735f5f947d9428680d8c095f6f70657261746f72948c076765746974656d9493948c06747970696e67948c04446963749493948c086275696c74696e73948c03737472949394682a8c03416e799493948694869452948c0672657475726e946834758c0e5f5f6b7764656661756c74735f5f944e8c0c5f5f64656661756c74735f5f944e8c0a5f5f6d6f64756c655f5f94681a8c075f5f646f635f5f944e8c0b5f5f636c6f737572655f5f944e8c175f636c6f75647069636b6c655f7375626d6f64756c6573945d948c0b5f5f676c6f62616c735f5f947d947586948652302e", "terminate_callback": "8005959f020000000000008c17636c6f75647069636b6c652e636c6f75647069636b6c65948c0e5f6d616b655f66756e6374696f6e9493942868008c0d5f6275696c74696e5f747970659493948c08436f6465547970659485945294284b034b004b004b034b054b83433e4b00010097007401000000000000000064017c006a020000000000000000000000000000000000009b0064029d03ab0100000000000001007900ad037701944e8c064167656e7420948c1520686173206265656e207465726d696e617465642e9487948c057072696e74948c046e616d659486948c056167656e74948c05696e707574948c066f75747075749487948c412f55736572732f617261747a2f666c6f636b2f6578616d706c65732f30325f636f6f6b5f626f6f6b2f736176655f616e645f6c6f61642f736176655f30322e7079948c0c6f6e5f7465726d696e617465948c0c6f6e5f7465726d696e617465944b0d431ce800f88000dc040988469035973a913a902cd01e33d00a34d50435f9944304821b1d01942929749452947d94288c0b5f5f7061636b6167655f5f948c00948c085f5f6e616d655f5f948c085f5f6d61696e5f5f948c085f5f66696c655f5f946813754e4e4e7494529468008c125f66756e6374696f6e5f736574737461746594939468217d947d9428681d8c0c6f6e5f7465726d696e617465948c0c5f5f7175616c6e616d655f5f948c0c6f6e5f7465726d696e617465948c0f5f5f616e6e6f746174696f6e735f5f947d948c0e5f5f6b7764656661756c74735f5f944e8c0c5f5f64656661756c74735f5f944e8c0a5f5f6d6f64756c655f5f94681e8c075f5f646f635f5f944e8c0b5f5f636c6f737572655f5f944e8c175f636c6f75647069636b6c655f7375626d6f64756c6573945d948c0b5f5f676c6f62616c735f5f947d947586948652302e", "on_error_callback": null}
