+++ examples/data/inherited_bloggy.json
@@ -1,0 +1,1 @@
+{"name": "inherited_bloggy", "model": "openai/gpt-4o", "description": "", "input": null, "output": null, "tools": null, "use_cache": true, "hand_off": null, "termination": null, "config": {"agent_type_override": null, "disable_output": false, "temperature": 0.0, "max_tokens": 2000}, "initialize_callback": null, "evaluate_callback": null, "terminate_callback": null, "on_error_callback": null}
